Address review feedback: move stubs to helpers and skip abort telemetry
Agent-Logs-Url: https://github.com/stellar/freighter/sessions/01d13312-6e53-44c0-936a-8b25cf011859
